+++ shared/util/github.go
@@ -7,10 +7,10 @@ import (
 	"strings"
 	"time"
 
+	"github.com/go-git/go-git/v5"
 	"github.com/machinebox/graphql"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
-	"gopkg.in/src-d/go-git.v4"
 )
 
 const GithubGraphQLEndpoint = "https://api.github.com/graphql"