+++ terraform/modules/happy-cloudfront/variables.tf
@@ -6,11 +6,18 @@ variable "frontend" {
   description = "The domain name and zone ID the user will see."
 }
 
-variable "origin" {
-  type = object({
-    domain_name = string
-  })
-  description = "The domain name of the origin."
+variable "origins" {
+  type = list(object({
+    domain_name              = string
+    path_pattern             = string
+    s3_origin_config         = optional(object({ origin_access_identity = string }))
+    origin_access_control_id = optional(string)
+  }))
+  description = "The domain names and the path used for the origin."
+  validation {
+    condition     = length(var.origins) > 0
+    error_message = "Must provide a least 1 origin"
+  }
 }
 
 variable "cache" {