+++ terraform/modules/happy-mesh-access-control/main.tf
@@ -1,6 +1,27 @@
 locals {
   allow_ingress_controller = var.service_type == "EXTERNAL" || var.service_type == "INTERNAL" || var.service_type == "VPC"
   needs_policy             = local.allow_ingress_controller || length(var.allow_mesh_services) > 0
+  # Service accounts that we want to allow access to this protected service
+  mesh_services_service_accounts = [for v in var.allow_mesh_services : {
+    "kind"      = "ServiceAccount"
+    "name"      = v.service_account_name != null && v.service_account_name != "" ? v.service_account_name : "${v.stack}-${v.service}-${var.deployment_stage}-${v.stack}"
+    "namespace" = var.k8s_namespace
+  }]
+  optional_ingress_controller_service_account = local.allow_ingress_controller ? [{
+    "kind"      = "ServiceAccount"
+    "name"      = "nginx-ingress-ingress-nginx"
+    "namespace" = "nginx-encrypted-ingress"
+  }] : []
+  status_page_service_account = [{
+    "kind"      = "ServiceAccount"
+    "name"      = "edu-platform-${var.deployment_stage}-status-page"
+    "namespace" = "status-page"
+  }]
+  k6_operator_service_account = var.allow_k6_operator ? [{
+    "kind"      = "ServiceAccount"
+    "name"      = "k6-operator-controller"
+    "namespace" = "k6-operator-system"
+  }] : []
 }
 
 resource "kubernetes_manifest" "linkerd_server" {
@@ -34,15 +55,12 @@ resource "kubernetes_manifest" "linkerd_mesh_tls_authentication" {
       "labels"    = var.labels
     }
     "spec" = {
-      "identityRefs" = concat([for v in var.allow_mesh_services : {
-        "kind"      = "ServiceAccount"
-        "name"      = v.service_account_name != null ? v.service_account_name : "${v.stack}-${v.service}-${var.deployment_stage}-${v.stack}"
-        "namespace" = var.k8s_namespace
-        }], local.allow_ingress_controller ? [{
-        "kind"      = "ServiceAccount"
-        "name"      = "nginx-ingress-ingress-nginx"
-        "namespace" = "nginx-encrypted-ingress"
-      }] : [])
+      "identityRefs" = concat(
+        local.mesh_services_service_accounts,
+        local.optional_ingress_controller_service_account,
+        local.status_page_service_account,
+        local.k6_operator_service_account
+      )
     }
   }
 }